feat: 시장 브리핑 데이터 정규화 및 AI 인프라 리팩토링
- 마켓 브리핑 정규화 기능 추가 (normalizePrimitives, 개선된 normalizeMarketBriefing)
- AI 인프라 계층 제거 (Claude, Gemini, analyzeAction API 관련 파일 삭제)
- 브리핑 카드 UI 및 레이아웃 개선
- 신뢰도 점수 계산 및 데이터 타입 정의 보완
- 테스트 파일 업데이트 및 통합
