--- src/__tests__/infrastructure/market/analyzeAction.test.ts
@@ -1,272 +1,0 @@
-import { analyzeAction } from '@/infrastructure/market/analyzeAction';
-import { EMPTY_SMC_RESULT } from '@/domain/indicators/constants';
-import type { Timeframe } from '@/domain/types';
-import type { RunAnalysisResult } from '@/infrastructure/market/analysisApi';
-
-jest.mock('@vercel/functions', () => ({
-    waitUntil: (promise: Promise<unknown>) => {
-        void promise;
-    },
-}));
-jest.mock('@/infrastructure/market/analysisApi');
-jest.mock('@/infrastructure/market/barsApi');
-jest.mock('@/infrastructure/cache/redis');
-jest.mock('@/infrastructure/cache/config', () => ({
-    ...jest.requireActual('@/infrastructure/cache/config'),
-    computeEffectiveTtl: jest.fn(),
-}));
-
-import { runAnalysis } from '@/infrastructure/market/analysisApi';
-import { fetchBarsWithIndicators } from '@/infrastructure/market/barsApi';
-import { createCacheProvider } from '@/infrastructure/cache/redis';
-import { computeEffectiveTtl } from '@/infrastructure/cache/config';
-
-const mockRunAnalysis = runAnalysis as jest.MockedFunction<typeof runAnalysis>;
-const mockFetchBarsWithIndicators =
-    fetchBarsWithIndicators as jest.MockedFunction<
-        typeof fetchBarsWithIndicators
-    >;
-const mockCreateCacheProvider = createCacheProvider as jest.MockedFunction<
-    typeof createCacheProvider
->;
-const mockComputeEffectiveTtl = computeEffectiveTtl as jest.MockedFunction<
-    typeof computeEffectiveTtl
->;
-
-const mockCacheGet = jest.fn();
-const mockCacheSet = jest.fn();
-const mockCacheDelete = jest.fn();
-
-const mockCacheProvider = {
-    get: mockCacheGet,
-    set: mockCacheSet,
-    delete: mockCacheDelete,
-};
-
-const mockSymbol = 'AAPL';
-const mockTimeframe: Timeframe = '1Day';
-
-const mockBarsData = {
-    bars: [
-        {
-            time: 1705312200,
-            open: 100,
-            high: 105,
-            low: 99,
-            close: 103,
-            volume: 1000000,
-        },
-    ],
-    indicators: {
-        macd: [],
-        bollinger: [],
-        dmi: [],
-        stochastic: [],
-        stochRsi: [],
-        rsi: [],
-        cci: [],
-        vwap: [],
-        ma: {},
-        ema: {},
-        volumeProfile: null,
-        ichimoku: [],
-        atr: [],
-        obv: [],
-        parabolicSar: [],
-        williamsR: [],
-        supertrend: [],
-        mfi: [],
-        keltnerChannel: [],
-        cmf: [],
-        donchianChannel: [],
-        buySellVolume: [],
-        squeezeMomentum: [],
-        smc: EMPTY_SMC_RESULT,
-    },
-};
-
-const mockResult: RunAnalysisResult = {
-    summary: '테스트 분석 요약',
-    trend: 'bullish',
-    indicatorResults: [],
-    riskLevel: 'low',
-    keyLevels: { support: [], resistance: [] },
-    priceTargets: {
-        bullish: { targets: [], condition: '' },
-        bearish: { targets: [], condition: '' },
-    },
-    patternSummaries: [],
-    strategyResults: [],
-    candlePatterns: [],
-    trendlines: [],
-    skillsDegraded: false,
-};
-
-describe('analyzeAction 함수는', () => {
-    beforeEach(() => {
-        mockRunAnalysis.mockReset();
-        mockFetchBarsWithIndicators.mockReset();
-        mockCacheGet.mockReset();
-        mockCacheSet.mockReset();
-        mockCacheDelete.mockReset();
-        mockComputeEffectiveTtl.mockReset();
-        mockCreateCacheProvider.mockReturnValue(mockCacheProvider);
-        mockFetchBarsWithIndicators.mockResolvedValue(mockBarsData);
-        // 기본 TTL: 1Day 기준 (86400)
-        mockComputeEffectiveTtl.mockReturnValue(86400);
-    });
-
-    describe('캐시 프로바이더가 없을 때', () => {
-        beforeEach(() => {
-            mockCreateCacheProvider.mockReturnValue(null);
-        });
-
-        it('fetchBarsWithIndicators를 호출하고 runAnalysis로 결과를 반환한다', async () => {
-            mockRunAnalysis.mockResolvedValueOnce(mockResult);
-
-            const result = await analyzeAction(mockSymbol, mockTimeframe);
-
-            expect(mockFetchBarsWithIndicators).toHaveBeenCalledWith(
-                mockSymbol,
-                mockTimeframe
-            );
-            expect(mockRunAnalysis).toHaveBeenCalledWith(
-                {
-                    symbol: mockSymbol,
-                    bars: mockBarsData.bars,
-                    indicators: mockBarsData.indicators,
-                },
-                mockTimeframe
-            );
-            expect(result).toBe(mockResult);
-        });
-    });
-
-    describe('캐시 히트일 때', () => {
-        it('fetchBarsWithIndicators와 runAnalysis를 호출하지 않고 캐시 결과를 반환한다', async () => {
-            mockCacheGet.mockResolvedValueOnce(mockResult);
-
-            const result = await analyzeAction(mockSymbol, mockTimeframe);
-
-            expect(mockCacheGet).toHaveBeenCalledWith('analysis:AAPL:1Day');
-            expect(mockFetchBarsWithIndicators).not.toHaveBeenCalled();
-            expect(mockRunAnalysis).not.toHaveBeenCalled();
-            expect(result).toBe(mockResult);
-        });
-    });
-
-    describe('캐시 미스일 때', () => {
-        it('runAnalysis를 호출하고 결과를 캐시에 저장한 뒤 반환한다', async () => {
-            mockCacheGet.mockResolvedValueOnce(null);
-            mockRunAnalysis.mockResolvedValueOnce(mockResult);
-            mockCacheSet.mockResolvedValueOnce(undefined);
-
-            const result = await analyzeAction(mockSymbol, mockTimeframe);
-
-            expect(mockRunAnalysis).toHaveBeenCalledWith(
-                {
-                    symbol: mockSymbol,
-                    bars: mockBarsData.bars,
-                    indicators: mockBarsData.indicators,
-                },
-                mockTimeframe
-            );
-            expect(result).toBe(mockResult);
-
-            // waitUntil로 등록된 백그라운드 작업이 실행되었는지 확인
-            await Promise.resolve();
-            expect(mockCacheSet).toHaveBeenCalledWith(
-                'analysis:AAPL:1Day',
-                mockResult,
-                86400
-            );
-        });
-
-        it('타임프레임에 맞는 TTL로 캐시를 저장한다', async () => {
-            mockCacheGet.mockResolvedValueOnce(null);
-            mockRunAnalysis.mockResolvedValueOnce(mockResult);
-            mockCacheSet.mockResolvedValueOnce(undefined);
-            mockComputeEffectiveTtl.mockReturnValueOnce(300);
-
-            await analyzeAction(mockSymbol, '5Min');
-
-            await Promise.resolve();
-            expect(mockCacheSet).toHaveBeenCalledWith(
-                'analysis:AAPL:5Min',
-                mockResult,
-                300
-            );
-        });
-    });
-
-    describe('캐시 읽기 에러일 때', () => {
-        it('에러를 로깅하고 runAnalysis를 실행한 뒤 결과를 캐시에 저장한다', async () => {
-            const consoleSpy = jest
-                .spyOn(console, 'error')
-                .mockImplementation(() => {});
-            const cacheError = new Error('Redis 연결 실패');
-            mockCacheGet.mockRejectedValueOnce(cacheError);
-            mockRunAnalysis.mockResolvedValueOnce(mockResult);
-            mockCacheSet.mockResolvedValueOnce(undefined);
-
-            const result = await analyzeAction(mockSymbol, mockTimeframe);
-
-            expect(consoleSpy).toHaveBeenCalledWith(
-                '[Cache] 캐시 읽기 실패:',
-                cacheError
-            );
-            expect(mockRunAnalysis).toHaveBeenCalledWith(
-                {
-                    symbol: mockSymbol,
-                    bars: mockBarsData.bars,
-                    indicators: mockBarsData.indicators,
-                },
-                mockTimeframe
-            );
-            expect(result).toBe(mockResult);
-
-            // 읽기 에러 후에도 waitUntil로 등록된 캐시 쓰기가 실행되는지 검증
-            await Promise.resolve();
-            expect(mockCacheSet).toHaveBeenCalledWith(
-                'analysis:AAPL:1Day',
-                mockResult,
-                86400
-            );
-            consoleSpy.mockRestore();
-        });
-    });
-
-    describe('캐시 쓰기 에러일 때', () => {
-        it('에러를 로깅하고 runAnalysis 결과를 반환한다', async () => {
-            const consoleSpy = jest
-                .spyOn(console, 'error')
-                .mockImplementation(() => {});
-            const cacheError = new Error('Redis 쓰기 실패');
-            mockCacheGet.mockResolvedValueOnce(null);
-            mockRunAnalysis.mockResolvedValueOnce(mockResult);
-            mockCacheSet.mockRejectedValueOnce(cacheError);
-
-            const result = await analyzeAction(mockSymbol, mockTimeframe);
-
-            expect(result).toBe(mockResult);
-
-            await Promise.resolve();
-            expect(consoleSpy).toHaveBeenCalledWith(
-                '[Cache] 캐시 쓰기 실패:',
-                cacheError
-            );
-            consoleSpy.mockRestore();
-        });
-    });
-
-    describe('runAnalysis가 에러를 던질 때', () => {
-        it('에러를 전파한다', async () => {
-            mockCacheGet.mockResolvedValueOnce(null);
-            mockRunAnalysis.mockRejectedValueOnce(new Error('Analysis failed'));
-
-            await expect(
-                analyzeAction(mockSymbol, mockTimeframe)
-            ).rejects.toThrow('Analysis failed');
-        });
-    });
-});
